+++ example/lib/page/home/view.dart
@@ -20,7 +20,7 @@ class HomePage extends StatelessWidget {
           title: Trs.appName.tr,
           backEnable: false,
           centerTitle: false,
-          systemOverlayStyle: getStatusBarStyle(isLightBarIcon: true),
+          systemOverlayStyle: getSystemOverlayStyle(sysNavigationBarColor: BaseColors.cPrimaryColor),
         ),
         body: SingleChildScrollView(
           child: Column(